moved _map_freq to utils.core
